simplify big endian compat
