+++ column_buffer.go
@@ -16,11 +16,7 @@ import (
 	"github.com/parquet-go/parquet-go/internal/bitpack"
 	"github.com/parquet-go/parquet-go/internal/unsafecast"
 	"github.com/parquet-go/parquet-go/sparse"
-	"golang.org/x/sys/cpu"
 )
-
-const offsetOfU64 = unsafe.Offsetof(Value{}.u64)
-const offsetOfPtr = unsafe.Offsetof(Value{}.ptr)
 
 // ColumnBuffer is an interface representing columns of a row group.
 //
@@ -105,29 +101,6 @@ func columnIndexOfNullable(base ColumnBuffer, maxDefinitionLevel byte, definitio
 		maxDefinitionLevel: maxDefinitionLevel,
 		definitionLevels:   definitionLevels,
 	}, nil
-}
-
-// On a big endian system, a boolean/byte value, which is in little endian byte format, is byte aligned
-// to the 7th byte in a u64 (8 bytes) variable.. Hence the data will be available at 7th byte when
-// interpreted as a little endian byte format. So, in order to access a boolean/byte value out of u64 variable,
-// we need to add an offset of "7"...
-// In the same way, an int32/uint32/float value, which is in little endian byte format, is byte aligned
-// to the 4th byte in a u64 (8 bytes) variable.. Hence the data will be available at 4th byte when
-// interpreted as a little endian byte format. So, in order to access an int32/uint32/float value out of u64 variable,
-// we need to add an offset of "4"
-func getOffset(colDict interface{}) uintptr {
-	var offset uintptr = 0
-
-	if cpu.IsBigEndian {
-		switch colDict.(type) {
-		case booleanColumnBuffer, booleanDictionary:
-			offset = 7
-
-		case int32ColumnBuffer, uint32ColumnBuffer, floatColumnBuffer, int32Dictionary, floatDictionary, uint32Dictionary:
-			offset = 4
-		}
-	}
-	return offset
 }
 
 type nullableColumnIndex struct {
@@ -855,8 +828,7 @@ func (col *booleanColumnBuffer) WriteBooleans(values []bool) (int, error) {
 }
 
 func (col *booleanColumnBuffer) WriteValues(values []Value) (int, error) {
-	offset := getOffset(*col)
-	col.writeValues(makeArrayValue(values, offsetOfU64+offset), columnLevels{})
+	col.writeValues(makeArrayValue(values, offsetOfBool), columnLevels{})
 	return len(values), nil
 }
 
@@ -995,8 +967,7 @@ func (col *int32ColumnBuffer) WriteInt32s(values []int32) (int, error) {
 }
 
 func (col *int32ColumnBuffer) WriteValues(values []Value) (int, error) {
-	offset := getOffset(*col)
-	col.writeValues(makeArrayValue(values, offsetOfU64+offset), columnLevels{})
+	col.writeValues(makeArrayValue(values, offsetOfU32), columnLevels{})
 	return len(values), nil
 }
 
@@ -1288,8 +1259,7 @@ func (col *floatColumnBuffer) WriteFloats(values []float32) (int, error) {
 }
 
 func (col *floatColumnBuffer) WriteValues(values []Value) (int, error) {
-	offset := getOffset(*col)
-	col.writeValues(makeArrayValue(values, offsetOfU64+offset), columnLevels{})
+	col.writeValues(makeArrayValue(values, offsetOfU32), columnLevels{})
 	return len(values), nil
 }
 
@@ -1776,8 +1746,7 @@ func (col *uint32ColumnBuffer) WriteUint32s(values []uint32) (int, error) {
 }
 
 func (col *uint32ColumnBuffer) WriteValues(values []Value) (int, error) {
-	offset := getOffset(*col)
-	col.writeValues(makeArrayValue(values, offsetOfU64+offset), columnLevels{})
+	col.writeValues(makeArrayValue(values, offsetOfU32), columnLevels{})
 	return len(values), nil
 }
