+++ value_le.go
@@ -1,0 +1,8 @@
+//go:build !s390x
+
+package parquet
+
+const (
+	firstByteOffsetOfBooleanValue = 0
+	firstByteOffsetOf32BitsValue  = 0
+)
